Menu/Settings + Loading Screen + Save/Load
Added a functioning menu that allows to adjust: music volume, graphical settings, fullscreen, screen resolution, v-sync, AA.

Through the menu you can quit the game with the quit button(only when the game is built).

A loading screen was added and the loading bar fills based on how fast the scene bakes and renders( almost instantly right now because it's a small scene ).

Save and Load system was added but only a basic one, it saves and loads the player's option settings after they click the apply button, the settings are stored in the company(Big Boy Produvtion)documents folder in %appdata% right now, the settings are saved even if the player leaves the game and rejoins it.