HealthBar/ Scripts + OcclusionCulling Added to UI Scene
Added a basic health bar with two scripts, so far it allows you to change the health bar values with the "Q" and "R" buttons, all the relavant values are public and easy to change.

Added Occlusion culling to my UI scene, it work for the most part but the occlusion culling can be seen in the minimap; as it's only affecting he main camera.

*Working on better UI designs and Concept Art.

*One more bug with the footsteps needs to get fixed.