UI updates and Colour-Blindness Settings
Changing the way the cameras render the minimap and shadows, the icons are also changed according to Darren's feedback.
Colour Blindness pack added but not working correctly, this will eventually help if we make these settings.
Occlusion still not working as intended, the minimap displays the occlusion happening behind the player.