Added ReadingBookScript & Fixed Issues with NPC
Basically added a script that rotates to the camera without it being in the NPC statement script. Added a WorldSpaceUIOverlay script so if you assign to a TMPro text it gets rendered first.
Added a script that allows the user to read a book with Interact "e" which is added onto inputmanager